refactor(check): Split typechecking tests into pass and fail
check/tests/typecheck.rs has gotten unorganized so at least splitting it up into passing and failing tests should bring a bit of order
